src/test/java/com/craig/learning/algorithm/PartitionLabelsTests.java
package com.craig.learning.algorithm;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.List;

import org.junit.jupiter.api.Test;

public class PartitionLabelsTests {
    PartitionLabels solution = new PartitionLabels();

    @Test
    void testPartitionLabels() {
        List<Integer> result = solution.partitionLabels("ababcbacadefegdehijhklij");
        assertEquals(3, result.size());
    } 
    
    @Test
    void testPartitionLabels2() {
        List<Integer> result = solution.partitionLabels("ntswuqqbidunnixxpoxxuuupotaatwdainsotwvpxpsdvdbwvbtdiptwtxnnbtqbdvnbowqitudutpsxsbbsvtipibqpvpnivottsxvoqqaqdxiviidivndvdtbvadnxboiqivpusuxaaqnqaobutdbpiosuitdnopoboivopaapadvqwwnnwvxndpxbapixaspwxxxvppoptqxitsvaaawxwaxtbxuixsoxoqdtopqqivaitnpvutzchkygjjgjkcfzjzrkmyerhgkglcyffezmehjcllmlrjghhfkfylkgyhyjfmljkzglkklykrjgrmzjyeyzrrkymccefggczrjflykclfhrjjckjlmglrmgfzlkkhffkjrkyfhegyykrzgjzcgjhkzzmzyejycfrkkekmhzjgggrmchkeclljlyhjkchmhjlehhejjyccyegzrcrerfzczfelzrlfylzleefgefgmzzlggmejjjygehmrczmkrc");
        assertEquals(2, result.size());
        assertEquals(246, result.get(0));
        assertEquals(254, result.get(1));
    }
}
